Retry connections in requests
