+++ src/project_config/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Project config utilities."""
